+++ recommends/storages/djangoorm/settings.py
@@ -1,0 +1,3 @@
+from django.conf import settings
+
+RECOMMENDS_STORAGE_DATABASE_ALIAS = getattr(settings, 'RECOMMENDS_STORAGE_DATABASE_ALIAS', 'recommends')