+++ docs/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 =========
 
 * v0.0.7
-    * added MongodbStorage
+    * added Mongodb storage
+    * added redis storage
 * v0.0.6
 	* added logging
 	* DjangoOrmStorage now saves Similarities and Suggestions in batches, according to the new ``RECOMMENDS_STORAGE_COMMIT_THRESHOLD`` setting.