+++ docs/changelog.rst
@@ -5,7 +5,7 @@ Changelog
 
 * v0.0.7
     * added Mongodb storage
-    * added redis storage
+    * added Redis storage
     * added ``unregister`` method to the registry
 * v0.0.6
 	* added logging